♻️ Refactor: 할일 인증 기능 패키지 구조 변경
