+++ feature/task-certification/src/main/java/com/twix/task_certification/certification/TaskCertificationViewModel.kt
@@ -1,11 +1,11 @@
-package com.twix.task_certification
+package com.twix.task_certification.certification
 
 import android.net.Uri
 import androidx.lifecycle.viewModelScope
-import com.twix.task_certification.model.TaskCertificationIntent
-import com.twix.task_certification.model.TaskCertificationSideEffect
-import com.twix.task_certification.model.TaskCertificationUiState
-import com.twix.task_certification.model.TorchStatus
+import com.twix.task_certification.certification.model.TaskCertificationIntent
+import com.twix.task_certification.certification.model.TaskCertificationSideEffect
+import com.twix.task_certification.certification.model.TaskCertificationUiState
+import com.twix.task_certification.certification.model.TorchStatus
 import com.twix.ui.base.BaseViewModel
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
