+++ feature/task-certification/src/main/java/com/twix/task_certification/certification/camera/Camera.kt
@@ -1,10 +1,10 @@
-package com.twix.task_certification.camera
+package com.twix.task_certification.certification.camera
 
 import android.net.Uri
 import androidx.camera.core.CameraSelector
 import androidx.lifecycle.LifecycleOwner
-import com.twix.task_certification.model.CameraPreview
-import com.twix.task_certification.model.TorchStatus
+import com.twix.task_certification.certification.model.CameraPreview
+import com.twix.task_certification.certification.model.TorchStatus
 import kotlinx.coroutines.flow.StateFlow
 
 interface Camera {
