+++ feature/task-certification/src/main/java/com/twix/task_certification/certification/component/TaskCertificationTopBar.kt
@@ -1,4 +1,4 @@
-package com.twix.task_certification.component
+package com.twix.task_certification.certification.component
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
