+++ feature/task-certification/src/main/java/com/twix/task_certification/certification/model/CameraPreview.kt
@@ -1,4 +1,4 @@
-package com.twix.task_certification.model
+package com.twix.task_certification.certification.model
 
 import androidx.camera.core.SurfaceRequest
 import androidx.compose.runtime.Immutable
