+++ feature/task-certification/src/main/java/com/twix/task_certification/certification/model/TaskCertificationSideEffect.kt
@@ -1,4 +1,4 @@
-package com.twix.task_certification.model
+package com.twix.task_certification.certification.model
 
 import com.twix.ui.base.SideEffect
