+++ feature/task-certification/src/main/java/com/twix/task_certification/certification/model/TorchStatus.kt
@@ -1,4 +1,4 @@
-package com.twix.task_certification.model
+package com.twix.task_certification.certification.model
 
 enum class TorchStatus {
     On,
