+++ feature/task-certification/src/main/java/com/twix/task_certification/di/TaskCertificationModule.kt
@@ -2,9 +2,9 @@ package com.twix.task_certification.di
 
 import com.twix.navigation.NavRoutes
 import com.twix.navigation.base.NavGraphContributor
-import com.twix.task_certification.TaskCertificationViewModel
-import com.twix.task_certification.camera.Camera
-import com.twix.task_certification.camera.CaptureCamera
+import com.twix.task_certification.certification.TaskCertificationViewModel
+import com.twix.task_certification.certification.camera.Camera
+import com.twix.task_certification.certification.camera.CaptureCamera
 import com.twix.task_certification.navigation.TaskCertificationGraph
 import org.koin.core.module.dsl.viewModelOf
 import org.koin.core.qualifier.named
